refactor: move `AiAssistantMessage` to design system
