+++ blocks/ai-chat/package.json
@@ -31,9 +31,7 @@
     "@blockprotocol/service": "0.1.3",
     "@hashintel/design-system": "0.0.6",
     "@mui/material": "5.11.8",
-    "lowlight": "2.8.1",
     "react-sizeme": "3.0.2",
-    "react-syntax-highlighter": "15.5.0",
     "react-transition-group": "4.4.5",
     "react-type-animation": "3.0.1",
     "uuid": "9.0.0"
@@ -42,7 +40,6 @@
     "@local/eslint-config": "0.0.0-private",
     "@local/tsconfig": "0.0.0-private",
     "@types/react-dom": "18.0.9",
-    "@types/react-syntax-highlighter": "15.5.6",
     "block-scripts": "0.2.3",
     "eslint": "8.33.0",
     "mock-block-dock": "0.1.3",
