+++ libs/@hashintel/design-system/package.json
@@ -28,7 +28,10 @@
     "@fortawesome/free-regular-svg-icons": "6.0.0",
     "@fortawesome/free-solid-svg-icons": "6.0.0",
     "clsx": "1.2.1",
-    "react-loading-skeleton": "3.2.0"
+    "lowlight": "2.8.1",
+    "react-loading-skeleton": "3.2.0",
+    "react-syntax-highlighter": "15.5.0",
+    "react-type-animation": "3.0.1"
   },
   "devDependencies": {
     "@babel/preset-env": "7.20.2",
@@ -44,6 +47,7 @@
     "@storybook/testing-library": "^0.0.14-next.1",
     "@types/react": "18.0.25",
     "@types/react-dom": "18.0.9",
+    "@types/react-syntax-highlighter": "15.5.6",
     "eslint": "8.33.0",
     "eslint-plugin-storybook": "0.6.10",
     "storybook": "7.0.0-beta.14",
