+++ libs/@hashintel/design-system/src/icons/icons.ts
@@ -1,5 +1,6 @@
 export { CaretDownIcon } from "./caret-down";
 export { CheckIcon } from "./check";
 export { ChevronRightIcon } from "./chevron-right";
+export { CodePenIcon } from "./code-pen";
 export { CopyIcon } from "./copy";
 export { GearIcon } from "./gear";
